Move substitute day helper functions under utils
- Moving these functions under utils because they will be used in common substitute operating period form
